major code cleanup & checking for admin rights

+++ main.py
@@ -21,8 +21,23 @@ bot = telebot.TeleBot(credentials["tg_token"])
 
 # подключить базу данных
 cluster = pymongo.MongoClient(
-    f"mongodb+srv://icton_bot:{credentials['mongodb_password']}@cluster0.yieez.mongodb.net/Icton?retryWrites=true&w=majority")
+    f"mongodb+srv://icton_bot:{credentials['mongodb_password']}@cluster0.yieez.mongodb.net/Icton?retryWrites=true&w"
+    f"=majority")
 db = cluster['Icton']
+
+
+class NoAdminRights(BaseException):
+    msg = "Ошибка: у пользователя нет прав администратора"
+
+
+# проверить пользователя на права администратора
+def check_user_for_admin_right(user, chat):
+    flag = False
+    for _user in bot.get_chat_administrators(chat.id):
+        if user.id == _user.user.id:
+            flag = True
+    if not flag:
+        raise NoAdminRights
 
 
 # получить текст сообщения из БД по его ID
@@ -31,25 +46,20 @@ def msg(db_msg_id, mongo_db=db):
     return messages.find_one({'_id': db_msg_id})['msg']
 
 
-# зарегистрировать пользователя в бд (и другая инициализация)
-@bot.message_handler(commands=['start'])
-def handle_user_reg(tg_message):
-    try:
-        user_tags.register_user(tg_message.from_user, db)
-    except user_tags.UserTagsException as e:
-        bot.send_message(tg_message.chat.id, e.msg)
-
-
 # прикрепить тэг к пользователю
 @bot.message_handler(commands=['add_tags'])
 def handle_tag_adding_start(tg_message):
-    response_msg = bot.send_message(tg_message.chat.id, "Укажи студента и тэг")
-    bot.register_next_step_handler(response_msg, handle_tag_adding_end)
-
-
-def handle_tag_adding_end(tg_message):
     try:
+        check_user_for_admin_right(tg_message.from_user, tg_message.chat)
 
+        response_msg = bot.send_message(tg_message.chat.id, "Укажи студента и тэг")
+        bot.register_next_step_handler(response_msg, handle_tag_adding_end, response_msg)
+    except NoAdminRights as e:
+        bot.send_message(tg_message.chat.id, e.msg)
+
+
+def handle_tag_adding_end(tg_message, info_msg):
+    try:
         username, tags = tg_message.text.split(' ', 1)
 
         if username == "" or tags == "":
@@ -60,6 +70,9 @@ def handle_tag_adding_end(tg_message):
         tags = tags.split(" ")
 
         user_tags.add_tags_to_user(username, tags, db)
+
+        bot.send_message(tg_message.chat.id,
+                         "Тэги " + " ".join(tags) + " успешно прикреплены к пользователю " + username)
     except user_tags.UserTagsException as e:
         bot.send_message(tg_message.chat.id, e.msg)
     except Exception:
@@ -68,53 +81,41 @@ def handle_tag_adding_end(tg_message):
 
 # уведомить пользователей с хотя бы одним из указанных тэгов
 @bot.message_handler(commands=['tag'])
-def handle_tag_start(tg_message):
-    if tg_message.reply_to_message is None:
-        bot.send_message(tg_message.chat.id, "Ответь этой командой на сообщение, под которым будут отмечены пользователи")
-        return
-
-    # удалить сообщение с командой, оно нам больше не нужно
-    bot.delete_message(tg_message.chat.id, tg_message.message_id)
-
-    # отправить подсказку и перенаправить ответ
-    response_msg = bot.send_message(tg_message.chat.id, "Укажи тэги")
-    bot.register_next_step_handler(response_msg, handle_tag_end, tg_message.reply_to_message, response_msg)
-
-
-def handle_tag_end(tg_message, original_message, info_message):
-    # удалить сообщение "Укажи тэги"
-    bot.delete_message(info_message.chat.id, info_message.message_id)
-
-    # обработать уведомление по его типу
-    handle_tag(original_message, tg_message, "")
+def handle_tag__start(tg_message):
+    handle_tag_middle(tg_message, "")
 
 
 # уведомить пользователей со всеми указанными тэгов
 @bot.message_handler(commands=['tag_all'])
 def handle_tag_all_start(tg_message):
-    if tg_message.reply_to_message is None:
-        bot.send_message(tg_message.chat.id, "Ответь этой командой на сообщение, под которым будут отмечены пользователи")
-        return
-
-    # удалить сообщение с командой, оно нам больше не нужно
-    bot.delete_message(tg_message.chat.id, tg_message.message_id)
-
-    # отправить подсказку и перенаправить ответ
-    response_msg = bot.send_message(tg_message.chat.id, "Укажи тэги")
-    bot.register_next_step_handler(response_msg, handle_tag_end, tg_message.reply_to_message, response_msg)
+    handle_tag_middle(tg_message, "all")
 
 
-def handle_tag_all_end(tg_message, original_message, info_message):
-    # удалить сообщение "Укажи тэги"
-    bot.delete_message(info_message.chat.id, info_message.message_id)
+def handle_tag_middle(tg_message, _type):
+    try:
+        check_user_for_admin_right(tg_message.from_user, tg_message.chat)
+        if tg_message.reply_to_message is None:
+            bot.send_message(tg_message.chat.id,
+                             "Ответь этой командой на сообщение, под которым будут отмечены пользователи")
+            return
 
-    # обработать уведомление по его типу
-    handle_tag(original_message, tg_message, "all")
+        # удалить сообщение с командой, оно нам больше не нужно
+        bot.delete_message(tg_message.chat.id, tg_message.message_id)
+
+        # отправить подсказку и перенаправить ответ
+        response_msg = bot.send_message(tg_message.chat.id, "Укажи тэги")
+        bot.register_next_step_handler(response_msg, handle_tag_end, tg_message.reply_to_message, response_msg, _type)
+
+    except NoAdminRights as e:
+        bot.send_message(tg_message.chat.id, e.msg)
 
 
 # главный обработчик уведомления пользователей по тэгам
-def handle_tag(original_message, tg_message, _type):
+def handle_tag_end(tg_message, original_message, info_message, _type):
     try:
+        # удалить сообщение "Укажи тэги"
+        bot.delete_message(info_message.chat.id, info_message.message_id)
+
         # распарсить тэги
         tags = tg_message.text.split(" ")
 
@@ -150,15 +151,28 @@ def handle_tag(original_message, tg_message, _type):
 # создать тэг в бд
 @bot.message_handler(commands=['create_tag'])
 def handle_tag_add_start(tg_message):
-    response_msg = bot.send_message(tg_message.chat.id, "Укажи название и описание тэга")
-    bot.register_next_step_handler(response_msg, handle_tag_add_end)
+    try:
+        check_user_for_admin_right(tg_message.from_user, tg_message.chat)
+
+        response_msg = bot.send_message(tg_message.chat.id, "Укажи название и описание тэга")
+        bot.register_next_step_handler(response_msg, handle_tag_add_end)
+
+    except NoAdminRights as e:
+        bot.send_message(tg_message.chat.id, e.msg)
 
 
 def handle_tag_add_end(tg_message):
     try:
-        tag_name, tag_description = tg_message.text.split(' ', 1)
+        index = tg_message.text.find(" ")
+        if index != -1:
+            tag_name, tag_description = tg_message.text[:index], tg_message.text[index:]
+        else:
+            tag_name = tg_message.text
+            tag_description = ""
 
         user_tags.add_tag(tag_name, tag_description, db)
+
+        bot.send_message(tg_message.chat.id, "Тэг " + tag_name + " успешно создан")
 
     except user_tags.UserTagsException as e:
         bot.send_message(tg_message.chat.id, e.msg)
@@ -184,8 +198,12 @@ def handle_get_all_tags(tg_message):
 # регистрация
 @bot.message_handler(commands=['register_me'])
 def handle(tg_message):
+    bot.delete_message(tg_message.chat.id, tg_message.message_id)
+
     try:
         user_tags.register_user(tg_message.from_user, db)
+        bot.send_message(tg_message.chat.id,
+                         "Пользователь " + tg_message.from_user.first_name + " успешно зарегестрирован.")
     except user_tags.UserTagsException as e:
         bot.send_message(tg_message.chat.id, e.msg)
 
@@ -227,42 +245,54 @@ def send_message(message):
 
 # добавить пункт в FAQ
 @bot.message_handler(commands=['faq_add'])
-def send_message(message):
-    text = 'В ответе на это сообщение пришлите текст нового пункта'
-    response_msg = bot.send_message(message.chat.id, text, parse_mode="html")
+def send_message(tg_message):
+    try:
+        check_user_for_admin_right(tg_message.from_user, tg_message.chat)
+        text = 'В ответе на это сообщение пришлите текст нового пункта'
+        response_msg = bot.send_message(tg_message.chat.id, text, parse_mode="html")
 
-    def _add_to_faq(new_text):
-        add_to_faq(db, new_text.text)
-        text = 'FAQ обновлен'
-        bot.send_message(message.chat.id, text, parse_mode="html")
+        def _add_to_faq(new_text):
+            add_to_faq(db, new_text.text)
+            text = 'FAQ обновлен'
+            bot.send_message(tg_message.chat.id, text, parse_mode="html")
 
-    bot.register_next_step_handler(response_msg, _add_to_faq)
+        bot.register_next_step_handler(response_msg, _add_to_faq)
+    except NoAdminRights as e:
+        bot.send_message(tg_message.chat.id, e.msg)
 
 
 # удалить пункт из FAQ
 @bot.message_handler(commands=['faq_remove'])
-def send_message(message):
-    text = 'В ответе на это сообщение пришлите номер удаляемого пункта'
-    response_msg = bot.send_message(message.chat.id, text, parse_mode="html")
+def send_message(tg_message):
+    try:
+        check_user_for_admin_right(tg_message.from_user, tg_message.chat)
+        text = 'В ответе на это сообщение пришлите номер удаляемого пункта'
+        response_msg = bot.send_message(tg_message.chat.id, text, parse_mode="html")
 
-    def _remove_from_faq(new_text):
+        def _remove_from_faq(new_text):
 
-        try:
-            remove_from_faq(db, int(new_text.text))
-            text = 'FAQ обновлен'
-        except ValueError:
-            text = 'Ошибка: ожидалось число. FAQ оставлен без изменений.'
-        bot.send_message(message.chat.id, text, parse_mode="html")
+            try:
+                remove_from_faq(db, int(new_text.text))
+                text = 'FAQ обновлен'
+            except ValueError:
+                text = 'Ошибка: ожидалось число. FAQ оставлен без изменений.'
+            bot.send_message(tg_message.chat.id, text, parse_mode="html")
 
-    bot.register_next_step_handler(response_msg, _remove_from_faq)
+        bot.register_next_step_handler(response_msg, _remove_from_faq)
+    except NoAdminRights as e:
+        bot.send_message(tg_message.chat.id, e.msg)
 
 
 # очистить FAQ
 @bot.message_handler(commands=['faq_flush'])
-def send_message(message):
-    flush_faq(db)
-    text = 'FAQ очищен'
-    bot.send_message(message.chat.id, text, parse_mode="html")
+def send_message(tg_message):
+    try:
+        check_user_for_admin_right(tg_message.from_user, tg_message.chat)
+        flush_faq(db)
+        text = 'FAQ очищен'
+        bot.send_message(tg_message.chat.id, text, parse_mode="html")
+    except NoAdminRights as e:
+        bot.send_message(tg_message.chat.id, e.msg)
 
 
 # получить информацию о преподавателях
@@ -270,7 +300,7 @@ def send_message(message):
 def get_teacher_info_prepare(message):
     _msg = bot.send_message(
         message.chat.id, 'Отправьте имя преподавателя.', parse_mode="html"
-        )
+    )
     bot.register_next_step_handler(_msg, get_teacher_info)
 
 
@@ -319,17 +349,17 @@ def send_message(message):
                  '/tomorrow - расписание на завтра',
                  '\n<b>Получение сведений</b>',
                  '/faq - вывести FAQ',
-                 '/faq_flush - очистить FAQ',
-                 '/faq_add - добавить пункт в FAQ',
-                 '/faq_remove - удалить пункт из FAQ',
+                 '/faq_flush - очистить FAQ - <b><u>адм.</u></b>',
+                 '/faq_add - добавить пункт в FAQ - <b><u>адм.</u></b>',
+                 '/faq_remove - удалить пункт из FAQ - <b><u>адм.</u></b>',
                  '/teacher - получить информацию о преподавателе',
                  '\n<b>Роли и тэги</b>',
                  '/register_me - регистрация в чате',
-                 '/create_tag - создать новый тэг',
+                 '/create_tag - создать новый тэг - <b><u>адм.</u></b>',
                  '/tags - получить список тэгов',
-                 '/add_tags - прикрепить к пользователю тэг',
-                 '/tag - уведомить всех пользователей с указанными тэгами',
-                 '/tag_all - уведомить только пользователей, имеющих все указанные теги',
+                 '/add_tags - прикрепить к пользователю тэг - <b><u>адм.</u></b>',
+                 '/tag - уведомить всех пользователей с указанными тэгами - <b><u>адм.</u></b>',
+                 '/tag_all - уведомить только пользователей, имеющих все указанные теги - <b><u>адм.</u></b>',
                  '\n<b>Прочее</b>',
                  '/help - помощь'
                  ]
